% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{LFS_RawData}
\alias{LFS_RawData}
\title{Labor Force Survey Raw Data}
\format{
A data frame with 101 rows and 97 variables:
\describe{
\item{column1}{Description of column1}
\item{column2}{Description of column2}
\item{...}{...}
}
}
\source{
Source of the data (e.g. "Collected from \link{amar.org}, original .mdb file")
}
\usage{
LFS_RawData
}
\description{
This data provides information about the Iranian labor market, including unemployment, employment, and participation rates.
}
\keyword{datasets}
